+++ tests/trace/test_trace_settings.py
@@ -3,12 +3,14 @@ import os
 import sys
 import time
 import timeit
+from unittest import mock
 
 import pytest
 
 import weave
 from weave.trace.constants import TRACE_CALL_EMOJI
 from weave.trace.settings import UserSettings, parse_and_apply_settings
+from weave.trace.weave_client import get_parallelism_settings
 
 
 @weave.op
@@ -163,21 +165,24 @@ def speed_test(client, count=5):
 
 
 def test_client_parallelism_setting(client_creator):
-    with client_creator() as client:
-        assert client.future_executor._max_workers == None
-        assert client.future_executor._executor._max_workers > 0
+    with mock.patch("os.cpu_count", return_value=4):
+        with client_creator() as client:
+            assert client.future_executor._max_workers == 4
+            assert client.future_executor._executor._max_workers == 4
 
     parse_and_apply_settings(UserSettings(client_parallelism=0))
-    with client_creator() as client:
-        assert client.future_executor._max_workers == 0
-        assert client.future_executor._executor == None
-        wait_time_0, queue_time_0 = speed_test(client)
+    with mock.patch("os.cpu_count", return_value=4):
+        with client_creator() as client:
+            assert client.future_executor._max_workers == 0
+            assert client.future_executor._executor == None
+            wait_time_0, queue_time_0 = speed_test(client)
 
     parse_and_apply_settings(UserSettings(client_parallelism=1))
-    with client_creator() as client:
-        assert client.future_executor._max_workers == 1
-        assert client.future_executor._executor._max_workers == 1
-        wait_time_1, queue_time_1 = speed_test(client)
+    with mock.patch("os.cpu_count", return_value=4):
+        with client_creator() as client:
+            assert client.future_executor._max_workers == 1
+            assert client.future_executor._executor._max_workers == 1
+            wait_time_1, queue_time_1 = speed_test(client)
 
     # Assert that the queue time is much less for 1 than 0
     assert queue_time_0 > queue_time_1
@@ -188,8 +193,9 @@ def test_client_parallelism_setting(client_creator):
 
     parse_and_apply_settings(UserSettings(client_parallelism=10))
     with client_creator() as client:
-        assert client.future_executor._max_workers == 10
-        assert client.future_executor._executor._max_workers == 10
+        assert client.future_executor._max_workers == 5
+        assert client.future_executor._executor._max_workers == 5
+        assert client.future_executor_fastlane._max_workers == 5
         wait_time_10, queue_time_10 = speed_test(client)
 
     # Assert that the queue time is about the same for 10 and 1
@@ -199,6 +205,60 @@ def test_client_parallelism_setting(client_creator):
 
     # Test explicit None
     parse_and_apply_settings(UserSettings(client_parallelism=None))
-    with client_creator() as client:
-        assert client.future_executor._max_workers == None
-        assert client.future_executor._executor._max_workers > 0
+    with mock.patch("os.cpu_count", return_value=4):
+        with client_creator() as client:
+            assert client.future_executor._max_workers == 4
+            assert client.future_executor._executor._max_workers > 0
+
+
+def test_get_parallelism_settings() -> None:
+    # Test default behavior with 4 CPU cores
+    with mock.patch("os.cpu_count", return_value=4):
+        main, upload = get_parallelism_settings()
+        assert main == 4  # (4 cores + 4) // 2 = 4
+        assert upload == 4  # Remaining threads for upload
+
+    # Test explicit total parallelism override
+    with mock.patch.dict(os.environ, {"WEAVE_CLIENT_PARALLELISM": "10"}):
+        main, upload = get_parallelism_settings()
+        assert main == 5  # 10 // 2 = 5
+        assert upload == 5  # Equal split
+
+    # Test explicit upload parallelism override
+    with mock.patch.dict(
+        os.environ,
+        {"WEAVE_CLIENT_PARALLELISM": "10", "WEAVE_CLIENT_PARALLELISM_UPLOAD": "3"},
+    ):
+        main, upload = get_parallelism_settings()
+        assert main == 10  # Total parallelism unchanged
+        assert upload == 3  # Explicit upload setting
+
+    # Test just upload parallelism override, 16 core machine
+    with mock.patch.dict(os.environ, {"WEAVE_CLIENT_PARALLELISM_UPLOAD": "3"}):
+        main, upload = get_parallelism_settings()
+        assert main is None  # will get set on thread pool init
+        assert upload == 3  # Explicit upload setting
+
+    # Test disabling parallelism
+    with mock.patch.dict(os.environ, {"WEAVE_CLIENT_PARALLELISM": "0"}):
+        main, upload = get_parallelism_settings()
+        assert main == 0
+        assert upload == 0
+
+    # Test max cap with many cores
+    with mock.patch("os.cpu_count", return_value=64):
+        main, upload = get_parallelism_settings()
+        assert main == 16  # (min(32, 68) // 2)
+        assert upload == 16  # Equal split of max 32
+
+    # Test single core system
+    with mock.patch("os.cpu_count", return_value=1):
+        main, upload = get_parallelism_settings()
+        assert main == 2  # (1 core + 4) // 2 = 2
+        assert upload == 3  # Remaining threads (5 - 2)
+
+    # Test when cpu_count returns None
+    with mock.patch("os.cpu_count", return_value=None):
+        main, upload = get_parallelism_settings()
+        assert main == 2  # (1 core + 4) // 2 = 2
+        assert upload == 3  # Remaining threads (5 - 2)
